+++ lib/base/base_game.dart
@@ -107,7 +107,7 @@ abstract class BaseGame extends FlameGame
 
   @override
   KeyEventResult onKeyEvent(
-    RawKeyEvent event,
+    KeyEvent event,
     Set<LogicalKeyboardKey> keysPressed,
   ) {
     if (!enabledKeyboard) {