+++ lib/input/keyboard/keyboard_listener.dart
@@ -4,7 +4,7 @@ import 'package:flutter/services.dart';
 mixin KeyboardEventListener on GameComponent {
   // If return 'true' this event is not relay to others components.
   bool onKeyboard(
-    RawKeyEvent event,
+    KeyEvent event,
     Set<LogicalKeyboardKey> keysPressed,
   ) {
     return false;