+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.6.0
+- Adds param `hudComponents` in `BonfireWidget`;
+- Adds `queryHud` method in `BonfireGameInterface`;
+- Adds `addHud` method in `BonfireGameInterface`;
+- Update Flame to `1.16.0`.
+
 # 3.5.0
 - Adds Parallax support
 - Background improvements.